Fixes for T3 v6.2
Configured constants to work with the constant editor with proper
explanations. Fixed path error for the email templates introduced with
changes to the template paths in setup.txt. Fixed syntax error with the
namespace in the Default.html layout. Changed the JS file adding a new
showAlert function that will allow you to output messages to the
colorbox form and added console logging to the ajax error function.