New Constants Editor Config and fixes
Added more configuration that can be done in the constants editor. Added
configuration for sending a notification email when a new user is
created. Renamed the create.json to Create.json so its properly found on
linux/unix hosts. Made a small change the postURL in the Cicregister.js
file while troubleshooting. Updated the plugin definitions in the
tables.php to use the vendor prefix in a standard way. Lots of changes
to the constants and setup typoscript files to allow more configuration
via the constants editor. Changed the SendValidationEmail behavior to
use a variable for the template file instead of it being hard-coded.
Modified the SendNotificationEmail behavior to user more variables.